+++ pkgs/native_assets_builder/test/build_runner/link_test.dart
@@ -26,15 +26,9 @@ void main() async {
           logger: logger,
         );
 
-        final buildResult = (await build(
+        final buildResult = (await buildDataAssets(
           packageUri,
-          logger,
-          dartExecutable,
           linkingEnabled: true,
-          buildAssetTypes: [DataAsset.type],
-          inputValidator: validateDataAssetBuildInput,
-          buildValidator: validateDataAssetBuildOutput,
-          applicationAssetValidator: (_) async => [],
         ))!;
         expect(buildResult.encodedAssets.length, 0);
 
@@ -50,15 +44,9 @@ void main() async {
         ))!;
         expect(linkResult.encodedAssets.length, 2);
 
-        final buildNoLinkResult = (await build(
+        final buildNoLinkResult = (await buildDataAssets(
           packageUri,
-          logger,
-          dartExecutable,
           linkingEnabled: false,
-          buildAssetTypes: [DataAsset.type],
-          inputValidator: validateDataAssetBuildInput,
-          buildValidator: validateDataAssetBuildOutput,
-          applicationAssetValidator: (_) async => [],
         ))!;
         expect(buildNoLinkResult.encodedAssets.length, 4);
       });
@@ -96,15 +84,9 @@ void main() async {
         // First, run `pub get`, we need pub to resolve our dependencies.
         await runPubGet(workingDirectory: packageUri, logger: logger);
 
-        final buildResult = await build(
+        final buildResult = await buildDataAssets(
           packageUri,
-          logger,
-          dartExecutable,
           linkingEnabled: true,
-          buildAssetTypes: [DataAsset.type],
-          inputValidator: validateDataAssetBuildInput,
-          buildValidator: validateDataAssetBuildOutput,
-          applicationAssetValidator: (_) async => [],
         );
         expect(buildResult, isNotNull);
         expect(_getNames(buildResult!.encodedAssets),
@@ -143,15 +125,9 @@ void main() async {
         logger: logger,
       );
 
-      final buildResult = (await build(
+      final buildResult = (await buildDataAssets(
         packageUri,
-        logger,
-        dartExecutable,
         linkingEnabled: true,
-        buildAssetTypes: [DataAsset.type],
-        inputValidator: validateDataAssetBuildInput,
-        buildValidator: validateDataAssetBuildOutput,
-        applicationAssetValidator: (_) async => [],
       ))!;
       expect(buildResult.encodedAssets.length, 0);
       expect(buildResult.encodedAssetsForLinking.length, 0);