+++ pkgs/native_assets_cli/test/code_assets/validation_test.dart
@@ -283,10 +283,10 @@ void main() {
       final errors =
           await validateCodeAssetBuildInput(BuildInput(builder.json));
 
-      bool matches(String error, String field) =>
-          RegExp('BuildInput.config.code.$field (.*foo baz).* does not'
-                  ' exist.')
-              .hasMatch(error);
+      bool matches(String error, String field) => RegExp(
+              'BuildInput.config.code.cCompiler.$field (.*foo baz).* does not'
+              ' exist.')
+          .hasMatch(error);
 
       expect(errors.any((e) => matches(e, 'compiler')), true);
       expect(errors.any((e) => matches(e, 'linker')), true);